Better tracing for WasmGoVM
