+++ packages/wasmvm/wasmhost/wasmcontext.go
@@ -20,6 +20,7 @@ const (
 
 type ISandbox interface {
 	Call(funcNr int32, params []byte) []byte
+	Tracef(format string, args ...interface{})
 }
 
 type WasmContext struct {
@@ -109,6 +110,9 @@ func (wc *WasmContext) callFunction() error {
 
 func (wc *WasmContext) ExportName(index int32, name string) {
 	if index >= 0 {
+		if HostTracing {
+			wc.tracef("ExportName(%d, %s)", index, name)
+		}
 		wc.funcTable.SetExport(index, name)
 		return
 	}
@@ -159,7 +163,14 @@ func (wc *WasmContext) log() iscp.LogInterface {
 }
 
 func (wc *WasmContext) Sandbox(funcNr int32, params []byte) []byte {
-	return wc.sandbox.Call(funcNr, params)
+	if !HostTracing || funcNr == wasmlib.FnLog || funcNr == wasmlib.FnTrace {
+		return wc.sandbox.Call(funcNr, params)
+	}
+
+	wc.tracef("Sandbox(%s)", traceSandbox(funcNr, params))
+	res := wc.sandbox.Call(funcNr, params)
+	wc.tracef("  => %s", hex(res))
+	return res
 }
 
 // state reduces the context state to a KVStoreReader
@@ -176,6 +187,9 @@ func (wc *WasmContext) state() kv.KVStoreReader {
 }
 
 func (wc *WasmContext) StateDelete(key []byte) {
+	if HostTracing {
+		wc.tracef("StateDelete(%s)", traceHex(key))
+	}
 	ctx := wc.wcSandbox.ctx
 	if ctx == nil {
 		panic("StateDelete: readonly state")
@@ -188,6 +202,9 @@ func (wc *WasmContext) StateExists(key []byte) bool {
 	if err != nil {
 		panic("StateExists: " + err.Error())
 	}
+	if HostTracing {
+		wc.tracef("StateExists(%s) = %v", traceHex(key), exists)
+	}
 	return exists
 }
 
@@ -196,13 +213,82 @@ func (wc *WasmContext) StateGet(key []byte) []byte {
 	if err != nil {
 		panic("StateGet: " + err.Error())
 	}
+	if HostTracing {
+		wc.tracef("StateGet(%s)", traceHex(key))
+		wc.tracef("  => %s", hex(res))
+	}
 	return res
 }
 
 func (wc *WasmContext) StateSet(key, value []byte) {
+	if HostTracing {
+		wc.tracef("StateSet(%s, %s)", traceHex(key), traceVal(value))
+	}
 	ctx := wc.wcSandbox.ctx
 	if ctx == nil {
 		panic("StateSet: readonly state")
 	}
 	ctx.State().Set(kv.Key(key), value)
+}
+
+func (wc *WasmContext) tracef(format string, args ...interface{}) {
+	if wc.proc != nil {
+		wc.log().Debugf(format, args...)
+		return
+	}
+	wc.sandbox.Tracef(format, args...)
+}
+
+func traceHex(key []byte) string {
+	name := ""
+	for i, b := range key {
+		if b == '.' {
+			return string(key[:i+1]) + hex(key[i+1:])
+		}
+		if b == '#' {
+			name = string(key[:i+1])
+			j := i + 1
+			for ; (key[j] & 0x80) != 0; j++ {
+			}
+			dec := wasmtypes.NewWasmDecoder(key[i+1 : j+1])
+			index := wasmtypes.Uint64Decode(dec)
+			name += wasmtypes.Uint64ToString(index)
+			if j+1 == len(key) {
+				return name
+			}
+			return name + "..." + hex(key[j+1:])
+		}
+	}
+	return `"` + string(key) + `"`
+}
+
+func traceSandbox(funcNr int32, params []byte) string {
+	name := sandboxFuncNames[-funcNr]
+	if name[0] == '$' {
+		return name[1:] + ", " + string(params)
+	}
+	if name[0] != '#' {
+		return name
+	}
+	return name[1:] + ", " + hex(params)
+}
+
+func traceVal(val []byte) string {
+	for _, b := range val {
+		if b < ' ' || b > '~' {
+			return hex(val)
+		}
+	}
+	return string(val)
+}
+
+// hex returns a hex string representing the byte buffer
+func hex(buf []byte) string {
+	const hexa = "0123456789abcdef"
+	res := make([]byte, len(buf)*2)
+	for i, b := range buf {
+		res[i*2] = hexa[b>>4]
+		res[i*2+1] = hexa[b&0x0f]
+	}
+	return string(res)
 }
