+++ packages/wasmvm/wasmlib/ts/wasmclient/configuration.ts
@@ -1,3 +1,6 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 import {Buffer} from './buffer';
 
 export interface IConfiguration {
